Переработал код
