+++ src/services/constants.js
@@ -2,5 +2,6 @@ const emotions = ['smile', 'sleeping', 'puke', 'angry'];
 const CARD_COUNT = 5;
 const CARD_TOP_COUNT = 2;
 const GENERATION_CARD_COUNT = 25;
+const COMMENT_COUNT = 75;
 
-export {emotions, CARD_COUNT, CARD_TOP_COUNT, GENERATION_CARD_COUNT};
+export {emotions, CARD_COUNT, CARD_TOP_COUNT, GENERATION_CARD_COUNT, COMMENT_COUNT};
