--- src/services/id-generation.js
@@ -1,9 +1,0 @@
-let idCount = 0;
-const generateId = (count) => {
-  for (let i = 0; i < count; i++) {
-    idCount++;
-    return idCount;
-  }
-};
-
-export {generateId};
