--- src/services/random-integer.js
@@ -1,6 +1,0 @@
-const getRandomInteger = (min, max) => {
-  const randomInteger = min - 0.5 + Math.random() * (max - min + 1);
-  return Math.round(randomInteger);
-};
-
-export {getRandomInteger};
